# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EMC - Physics NEET Prep API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Topic {
  id: number;
  name: string;
  description: string;
  questionCount: number;
  icon: string;
}

export type QuestionDifficulty =
  (typeof QuestionDifficulty)[keyof typeof QuestionDifficulty];

export const QuestionDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;

export interface Question {
  id: number;
  topicId: number;
  topicName: string;
  text: string;
  options: string[];
  correctOption: number;
  explanation: string;
  difficulty: QuestionDifficulty;
  year?: number | null;
}

export type MockTestDifficulty =
  (typeof MockTestDifficulty)[keyof typeof MockTestDifficulty];

export const MockTestDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
  mixed: "mixed",
} as const;

export interface MockTest {
  id: number;
  title: string;
  description: string;
  questionCount: number;
  durationMinutes: number;
  difficulty: MockTestDifficulty;
  topicIds: number[];
}

export type MockTestWithQuestionsDifficulty =
  (typeof MockTestWithQuestionsDifficulty)[keyof typeof MockTestWithQuestionsDifficulty];

export const MockTestWithQuestionsDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
  mixed: "mixed",
} as const;

export interface MockTestWithQuestions {
  id: number;
  title: string;
  description: string;
  durationMinutes: number;
  difficulty: MockTestWithQuestionsDifficulty;
  questions: Question[];
}

export type CreateTestBodyDifficulty =
  (typeof CreateTestBodyDifficulty)[keyof typeof CreateTestBodyDifficulty];

export const CreateTestBodyDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
  mixed: "mixed",
} as const;

export interface CreateTestBody {
  title: string;
  description?: string;
  topicIds: number[];
  questionCount: number;
  durationMinutes: number;
  difficulty: CreateTestBodyDifficulty;
}

export interface AnswerInput {
  questionId: number;
  selectedOption: number | null;
}

export interface SubmitAttemptBody {
  testId: number;
  timeTakenSeconds: number;
  answers: AnswerInput[];
}

export interface Attempt {
  id: number;
  testId: number;
  testTitle: string;
  score: number;
  totalQuestions: number;
  correctAnswers: number;
  timeTakenSeconds: number;
  completedAt: string;
}

export interface AnswerDetail {
  questionId: number;
  questionText: string;
  selectedOption: number | null;
  correctOption: number;
  isCorrect: boolean;
  explanation: string;
  topicName: string;
}

export interface AttemptResult {
  id: number;
  testId: number;
  testTitle: string;
  score: number;
  totalQuestions: number;
  correctAnswers: number;
  timeTakenSeconds: number;
  completedAt: string;
  answers: AnswerDetail[];
}

export interface ProgressSummary {
  totalQuestionsAnswered: number;
  totalCorrect: number;
  accuracy: number;
  totalTestsTaken: number;
  averageScore: number;
  bestScore: number;
  streak: number;
  weakTopics: string[];
  strongTopics: string[];
}

export interface TopicProgress {
  topicId: number;
  topicName: string;
  questionsAttempted: number;
  correctAnswers: number;
  accuracy: number;
  totalQuestions: number;
}

export type ListQuestionsParams = {
  topicId?: number;
  difficulty?: ListQuestionsDifficulty;
  limit?: number;
};

export type ListQuestionsDifficulty =
  (typeof ListQuestionsDifficulty)[keyof typeof ListQuestionsDifficulty];

export const ListQuestionsDifficulty = {
  easy: "easy",
  medium: "medium",
  hard: "hard",
} as const;