# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EMC - Physics NEET Prep API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all physics topics
 */
export const ListTopicsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  questionCount: zod.number(),
  icon: zod.string(),
});
export const ListTopicsResponse = zod.array(ListTopicsResponseItem);

/**
 * @summary Get a single topic
 */
export const GetTopicParams = zod.object({
  id: zod.coerce.number(),
});

export const GetTopicResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  questionCount: zod.number(),
  icon: zod.string(),
});

/**
 * @summary List questions with optional filters
 */
export const ListQuestionsQueryParams = zod.object({
  topicId: zod.coerce.number().optional(),
  difficulty: zod.enum(["easy", "medium", "hard"]).optional(),
  limit: zod.coerce.number().optional(),
});

export const ListQuestionsResponseItem = zod.object({
  id: zod.number(),
  topicId: zod.number(),
  topicName: zod.string(),
  text: zod.string(),
  options: zod.array(zod.string()),
  correctOption: zod.number(),
  explanation: zod.string(),
  difficulty: zod.enum(["easy", "medium", "hard"]),
  year: zod.number().nullish(),
});
export const ListQuestionsResponse = zod.array(ListQuestionsResponseItem);

/**
 * @summary Get a single question with explanation
 */
export const GetQuestionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetQuestionResponse = zod.object({
  id: zod.number(),
  topicId: zod.number(),
  topicName: zod.string(),
  text: zod.string(),
  options: zod.array(zod.string()),
  correctOption: zod.number(),
  explanation: zod.string(),
  difficulty: zod.enum(["easy", "medium", "hard"]),
  year: zod.number().nullish(),
});

/**
 * @summary List available mock tests
 */
export const ListTestsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  questionCount: zod.number(),
  durationMinutes: zod.number(),
  difficulty: zod.enum(["easy", "medium", "hard", "mixed"]),
  topicIds: zod.array(zod.number()),
});
export const ListTestsResponse = zod.array(ListTestsResponseItem);

/**
 * @summary Create a new mock test
 */
export const CreateTestBody = zod.object({
  title: zod.string(),
  description: zod.string().optional(),
  topicIds: zod.array(zod.number()),
  questionCount: zod.number(),
  durationMinutes: zod.number(),
  difficulty: zod.enum(["easy", "medium", "hard", "mixed"]),
});

/**
 * @summary Get a mock test with its questions
 */
export const GetTestParams = zod.object({
  id: zod.coerce.number(),
});

export const GetTestResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  durationMinutes: zod.number(),
  difficulty: zod.enum(["easy", "medium", "hard", "mixed"]),
  questions: zod.array(
    zod.object({
      id: zod.number(),
      topicId: zod.number(),
      topicName: zod.string(),
      text: zod.string(),
      options: zod.array(zod.string()),
      correctOption: zod.number(),
      explanation: zod.string(),
      difficulty: zod.enum(["easy", "medium", "hard"]),
      year: zod.number().nullish(),
    }),
  ),
});

/**
 * @summary List all past attempts
 */
export const ListAttemptsResponseItem = zod.object({
  id: zod.number(),
  testId: zod.number(),
  testTitle: zod.string(),
  score: zod.number(),
  totalQuestions: zod.number(),
  correctAnswers: zod.number(),
  timeTakenSeconds: zod.number(),
  completedAt: zod.string(),
});
export const ListAttemptsResponse = zod.array(ListAttemptsResponseItem);

/**
 * @summary Submit a completed test attempt
 */
export const SubmitAttemptBody = zod.object({
  testId: zod.number(),
  timeTakenSeconds: zod.number(),
  answers: zod.array(
    zod.object({
      questionId: zod.number(),
      selectedOption: zod.number().nullable(),
    }),
  ),
});

/**
 * @summary Get a specific attempt with detailed results
 */
export const GetAttemptParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAttemptResponse = zod.object({
  id: zod.number(),
  testId: zod.number(),
  testTitle: zod.string(),
  score: zod.number(),
  totalQuestions: zod.number(),
  correctAnswers: zod.number(),
  timeTakenSeconds: zod.number(),
  completedAt: zod.string(),
  answers: zod.array(
    zod.object({
      questionId: zod.number(),
      questionText: zod.string(),
      selectedOption: zod.number().nullable(),
      correctOption: zod.number(),
      isCorrect: zod.boolean(),
      explanation: zod.string(),
      topicName: zod.string(),
    }),
  ),
});

/**
 * @summary Get overall progress summary (dashboard stats)
 */
export const GetProgressSummaryResponse = zod.object({
  totalQuestionsAnswered: zod.number(),
  totalCorrect: zod.number(),
  accuracy: zod.number(),
  totalTestsTaken: zod.number(),
  averageScore: zod.number(),
  bestScore: zod.number(),
  streak: zod.number(),
  weakTopics: zod.array(zod.string()),
  strongTopics: zod.array(zod.string()),
});

/**
 * @summary Get topic-wise performance breakdown
 */
export const GetTopicProgressResponseItem = zod.object({
  topicId: zod.number(),
  topicName: zod.string(),
  questionsAttempted: zod.number(),
  correctAnswers: zod.number(),
  accuracy: zod.number(),
  totalQuestions: zod.number(),
});
export const GetTopicProgressResponse = zod.array(GetTopicProgressResponseItem);